Rename Rio to Jarvis throughout project
All references to 'Rio' have been renamed to 'Jarvis' in source code, documentation, configuration, and data files. This includes package names, class names, file names, and user-facing strings to reflect the new project identity.
