+++ docs/README.md
@@ -1,4 +1,4 @@
-# rio.Rio User Guide
+# jarvis.Jarvis User Guide
 
 // Update the title above to match the actual product name
