+++ src/main/java/jarvis/command/TodoCommand.java
@@ -1,8 +1,8 @@
-package rio.command;
+package jarvis.command;
 
-import rio.TaskList;
-import rio.task.Task;
-import rio.task.ToDo;
+import jarvis.TaskList;
+import jarvis.task.Task;
+import jarvis.task.ToDo;
 
 /**
  * Represents a command that creates a t-odo task and adds it to the
