+++ src/main/java/jarvis/task/Deadline.java
@@ -1,4 +1,4 @@
-package rio.task;
+package jarvis.task;
 
 /**
  * Represents a task with a specific deadline.
