+++ src/main/java/jarvis/task/Event.java
@@ -1,4 +1,4 @@
-package rio.task;
+package jarvis.task;
 
 /**
  * Represents a event task.
