+++ src/main/java/jarvis/task/ToDo.java
@@ -1,4 +1,4 @@
-package rio.task;
+package jarvis.task;
 
 /**
  * Represents a normal task.
