+++ src/test/java/jarvis/task/DeadlineTest.java
@@ -1,4 +1,4 @@
-package rio.task;
+package jarvis.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
