+++ src/test/java/jarvis/task/ToDoTest.java
@@ -1,4 +1,4 @@
-package rio.task;
+package jarvis.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
