+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
    ____________________________________________________________
-    Hello, I'm rio.Rio - your personal assistant.
+    Hello, I'm jarvis.Jarvis - your personal assistant.
     What can I do for you today?
    ____________________________________________________________
